refactor hook that doesn't need to be a hook
